+++ example/lib/main.dart
@@ -1,7 +1,4 @@
-import 'package:bangla_converter/model/bangla_date_return_model.dart';
 import 'package:flutter/material.dart';
-
-///import [bangla_converter] package
 import 'package:bangla_converter/bangla_converter.dart';
 
 void main() {
@@ -11,11 +8,10 @@ void main() {
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Bangla Convertor',
+      title: 'Bangla Converter',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
@@ -61,12 +57,11 @@ class _MyHomePageState extends State<MyHomePage> {
           ),
         ),
       ),
-
       floatingActionButton: FloatingActionButton(
         onPressed: _incrementCounter,
         tooltip: 'Increment',
         child: const Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
+      ),
     );
   }
 
@@ -100,14 +95,14 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   Column _dateConvert() {
-    BanglaDateReturn? _banglaToDate = BanglaConverter.banglaDate(isToday: true);
-    BanglaDateReturn? _banglaDateByDate = BanglaConverter.banglaDate(
+    BanglaDateReturn? banglaToDate = BanglaConverter.banglaDate(isToday: true);
+    BanglaDateReturn? banglaDateByDate = BanglaConverter.banglaDate(
       englishDate: DateTime.parse("2023-02-01"),
     );
-    BanglaDateReturn? _banglaDateByValue = BanglaConverter.banglaDate(
-      day: "14",
-      month: "4",
-      year: "2024",
+    BanglaDateReturn? banglaDateByValue = BanglaConverter.banglaDate(
+      day: 14,
+      month: 4,
+      year: 2024,
     );
     return Column(
       children: [
@@ -122,7 +117,7 @@ class _MyHomePageState extends State<MyHomePage> {
           style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
         ),
         Text(
-          '${_banglaToDate!.date} ${_banglaToDate.month} ${_banglaToDate.year} বঙ্গাব্দ, ${_banglaToDate.weekDay}',
+          '${banglaToDate!.date} ${banglaToDate.month} ${banglaToDate.year} বঙ্গাব্দ, ${banglaToDate.weekDay}',
           style: Theme.of(context).textTheme.headlineMedium,
         ),
         SizedBox(height: 15),
@@ -132,28 +127,27 @@ class _MyHomePageState extends State<MyHomePage> {
           style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
         ),
         Text(
-          '${_banglaDateByDate!.date} ${_banglaDateByDate.month} ${_banglaDateByDate.year} বঙ্গাব্দ, ${_banglaDateByDate.weekDay}',
+          '${banglaDateByDate!.date} ${banglaDateByDate.month} ${banglaDateByDate.year} বঙ্গাব্দ, ${banglaDateByDate.weekDay}',
           style: Theme.of(context).textTheme.headlineMedium,
         ),
         SizedBox(height: 15),
         const Text(
-          'Bangla Date by Date, Month, Year \n EX: BanglaConverter.banglaDate(day: "5",month: "4",year: "2024")',
+          'Bangla Date by Date, Month, Year \n EX: BanglaConverter.banglaDate(day: 14, month: 4, year: 2024)',
           textAlign: TextAlign.center,
           style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
         ),
         Text(
-          '${_banglaDateByValue!.date} ${_banglaDateByValue.month} ${_banglaDateByValue.year} বঙ্গাব্দ, ${_banglaDateByValue.weekDay}',
+          '${banglaDateByValue!.date} ${banglaDateByValue.month} ${banglaDateByValue.year} বঙ্গাব্দ, ${banglaDateByValue.weekDay}',
           style: Theme.of(context).textTheme.headlineMedium,
         ),
-
         SizedBox(height: 15),
         const Text(
-          'English Date by Bangle Date, Month, Year \n EX: BanglaConverter.englishDate(banglaDay: "1", banglaMonth: "1", banglaYear: "1432")',
+          'English Date by Bangla Date, Month, Year \n EX: BanglaConverter.englishDate(banglaDay: 1, banglaMonth: 1, banglaYear: 1432)',
           textAlign: TextAlign.center,
           style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
         ),
         Text(
-          '${BanglaConverter.englishDate(banglaDay: "1", banglaMonth: "1", banglaYear: "1432")}',
+          '${BanglaConverter.englishDate(banglaDay: 1, banglaMonth: 1, banglaYear: 1432)}',
           style: Theme.of(context).textTheme.headlineMedium,
         ),
       ],